remove old code and restructure hud / entity rendering
